requirements.txt
amqp==2.2.1
Babel==2.4.0
billiard==3.5.0.3
celery==4.1.0
certifi==2017.4.17
chardet==3.0.4
dj-database-url==0.4.2
Django==1.11.3
django-paypal==0.3.6
flower==0.9.2
gunicorn==19.7.1
idna==2.5
imgurpython==1.1.7
kombu==4.1.0
numpy==1.12.0
olefile==0.44
Pillow==4.2.1
psycopg2==2.7.3
pytz==2017.2
requests==2.18.1
six==1.10.0
tornado==4.5.1
urllib3==1.21.1
vine==1.1.4
virtualenv==15.1.0
whitenoise==3.3.0
